TASK: Move `NodeAggregateIdMapping` to Neos as it will be removed from the cr core soon
